requirements.txt
re
os
csv
sqlalchemy
sqlparse-0.4.4
tkinter
